LayoutTests/fast/history/gesture-before-onload-expected.txt

This page is the target of a redirect.

PASS: History item count should be 2 and is.


============== Back Forward List ==============
        (file test):fast/history/gesture-before-onload.html  **nav target**
curr->  (file test):fast/history/resources/gesture-before-onload-target.html#2  **nav target**
===============================================
